Hotfixs/ricardo

See merge request lbaw/lbaw2223/lbaw2261!38

+++ app/Http/Controllers/CommentController.php
@@ -8,14 +8,16 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Events\NewNotification;
+use App\Models\Post;
 
 class CommentController extends Controller
 {
     public function create($id_post, Request $request)
     {
-        // $this->authorize('create', Comment::class);
-
         $text = $request->input('text');
+
+        $post = Post::find($id_post);
+        $this->authorize('create', $post); // funciona
 
         preg_match_all('/(?<=@)\w+/m', $text, $matches);
 
@@ -79,7 +81,7 @@ class CommentController extends Controller
     public function edit(Request $request)
     {
         $comment = Comment::find($request->input('id_comment'));
-        //$this->authorize('edit', $comment);
+        $this->authorize('update', $comment);
         $comment->text = $request->input('text');
         $comment->save();
     }
@@ -87,7 +89,7 @@ class CommentController extends Controller
     public function delete($id_comment)
     {
         $comment = Comment::find($id_comment);
-        //$this->authorize('delete', $comment);
+        $this->authorize('delete', $comment);
         $comment->delete();
     }
 }
